Return vector of uint probabilities for each vertex

+++ libs/mvs_tex_wrapper/wrapper.h
@@ -35,5 +35,5 @@ void textureMesh(const TextureSettings& texture_settings,
                  std::shared_ptr<EuclideanViewMask> ev_mask = NULL,
                  uint atlas_size = 0,
                  float* hidden_face_proportion = NULL,
-                 std::vector<std::vector<int>> *segmentation_classes = nullptr);
+                 std::vector<std::vector<uint8_t>> *segmentation_classes = nullptr);
 }  // namespace MvsTexturing
